+++ src/main/java/telraam/App.java
@@ -17,9 +17,9 @@ import telraam.api.*;
 import telraam.database.daos.*;
 import telraam.database.models.Station;
 import telraam.healthchecks.TemplateHealthCheck;
-import telraam.logic.Lapper;
-import telraam.logic.external.ExternalLapper;
-import telraam.logic.robust.RobustLapper;
+import telraam.logic.lapper.Lapper;
+import telraam.logic.lapper.external.ExternalLapper;
+import telraam.logic.lapper.robust.RobustLapper;
 import telraam.station.Fetcher;
 import telraam.util.AcceptedLapsUtil;
 import telraam.websocket.WebSocketConnection;
