+++ src/main/java/telraam/logic/lapper/Lapper.java
@@ -1,4 +1,4 @@
-package telraam.logic;
+package telraam.logic.lapper;
 
 import io.dropwizard.jersey.setup.JerseyEnvironment;
 import telraam.database.models.Detection;
