+++ src/main/java/telraam/logic/lapper/external/ExternalLapper.java
@@ -1,4 +1,4 @@
-package telraam.logic.external;
+package telraam.logic.lapper.external;
 
 import io.dropwizard.jersey.setup.JerseyEnvironment;
 import org.jdbi.v3.core.Jdbi;
@@ -7,8 +7,8 @@ import telraam.database.daos.LapSourceDAO;
 import telraam.database.models.Detection;
 import telraam.database.models.Lap;
 import telraam.database.models.LapSource;
-import telraam.logic.Lapper;
-import telraam.logic.external.models.ExternalLapperTeamLaps;
+import telraam.logic.lapper.Lapper;
+import telraam.logic.lapper.external.models.ExternalLapperTeamLaps;
 
 import java.sql.Timestamp;
 import java.util.Comparator;
