+++ src/main/java/telraam/station/Fetcher.java
@@ -7,7 +7,7 @@ import telraam.database.daos.StationDAO;
 import telraam.database.models.Baton;
 import telraam.database.models.Detection;
 import telraam.database.models.Station;
-import telraam.logic.Lapper;
+import telraam.logic.lapper.Lapper;
 import telraam.station.models.RonnyDetection;
 import telraam.station.models.RonnyResponse;
