+++ src/test/java/telraam/logic/simple/SimpleLapperTest.java
@@ -8,8 +8,8 @@ import telraam.database.daos.LapSourceDAO;
 import telraam.database.models.Detection;
 import telraam.database.models.Lap;
 import telraam.database.models.LapSource;
-import telraam.logic.Lapper;
-import telraam.logic.simple.SimpleLapper;
+import telraam.logic.lapper.Lapper;
+import telraam.logic.lapper.simple.SimpleLapper;
 
 import java.sql.Timestamp;
 import java.util.Optional;
